pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "evidence-based-hdb-resale-market-analysis"
version = "0.1.0"
description = "Reproducible HDB resale case-study pipeline, analytics, and deck artifacts."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "catboost>=1.2",
  "kaleido>=0.2.1,<1",
  "matplotlib>=3.10",
  "numpy>=2.2",
  "pandas>=2.3",
  "plotly>=6.0",
  "pyarrow>=23.0",
  "python-pptx>=1.0",
  "PyYAML>=6.0",
  "requests>=2.32",
  "scikit-learn>=1.7",
  "scipy>=1.15",
  "seaborn>=0.13",
  "statsmodels>=0.14",
  "tqdm>=4.67",
  "xgboost>=3.2",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
